nxp-ls1046: remove u-boot-qoriq package from this BSP
Issue: LINUXEXEC-19886

For nxp-ls1046 BSP, because of the license issue, the whole bootloader
is not able to be built in wrlinux product directly and user need to download
the official NXP bootloader for ls1046rdba board. In this way, it doesn't
have any meaning to keep the only u-boot package in this BSP. So, remove it
to avoid confusing customer.

Signed-off-by: Meng Li <Meng.Li@windriver.com>